Add defmodule
This is what we aim to create:
```
defmodule ModuleName do
  def function_name
    ...
  end
end
```

The first step is to open up object and create a method:
```
class Object
  def defmodule(name, &block)
  end
end
```

We want to pass in a name and a block. If we try and pass in a constant, e.g. `MyModule`,
then we will get `NameError (uninitialized constant MyModule)`
So we need a way for Ruby to catch this. We can open up Object and redefine `self.const_missing`
```
class Object
  def self.const_missing(name)
    name
  end
end
```

This returns the constant as a Symbol.
```
irb(main):001:0> HelloWorld
=> :HelloWorld
```
This is is good because this is also what Elixir does. In Elixir module names are Atoms.

We can now create an empty module with a name that we pass in. Here name is a Symbol as returned by const_missing so we have to turn it into a string.
Here we are writing Ruby code and evaluating it with eval:
```
eval("#{name.to_s} = Module.new()")
```

We can use:
```
m = Kernel.const_get(name.to_s)
```
to get the actual constant. Then we can call `module_eval` and pass in the block unmodified.

Now we have a module that has instance methods defined. We need to make the instance methods module functions.
This normally looks like module_function :first_method, :second_method.

We can call `instance_methods` on our new module to get a list of all the methods we want to make module_functions. Then we can open it up with eval and run module_function inside it. We have to convert the array to a string for eval so that it is valid Ruby code.