Implement def ... do inside defmodule
First we update the new module to return a recursive syntax tree with method_missing
```
m.class_eval do
      def method_missing(m_missing_name, *m_missing_args, &m_missing_block)
        [m_missing_name, m_missing_args, m_missing_block]
      end

      module_function :method_missing
    end
```

This will return a tree in the form of [name, args, block] where args will look like [[name, [], block]]

Then we store some state inside the class variables.

To track the list of defined functions
```
    m.class_variable_set(:@@public_functions, {})
```

To store the context of the current evaluation of the module.
This is necessary because method missing will recurse and we need to keep the context somewhere.
```
    m.class_variable_set(:@@current_arg_values, [])
    m.class_variable_set(:@@current_arg_names, [])
```

Then we define defP. This stores a list of defined functions on the class.
defP short for defPublic. Not strictly Elixiry but we can't override def.
```
      def defP(ast, &block)
        function_name = ast[0]
        arg_names = ast[1].map {|list| list[0]}
        arity = arg_names.size

        p = self.class_variable_get(:@@public_functions)
        p[{name: function_name, arity: arity}] = {args: arg_names, proc: block}
        p = self.class_variable_set(:@@public_functions, p)
      end
```

And we redefine method_missing which will handle calls to the function and variables by calling method_missing or raise an error if it cannot find anything
```
      m.class_eval do
        def method_missing(m_missing_name, *m_missing_args, &m_missing_block)

          if f = self.class_variable_get(:@@public_functions)[{name: m_missing_name, arity: m_missing_args.size}]
            self.class_variable_set(:@@current_arg_values, m_missing_args)
            self.class_variable_set(:@@current_arg_names, f[:args])

            f[:proc].call
          elsif index = self.class_variable_get(:@@current_arg_names).find_index(m_missing_name)
            self.class_variable_get(:@@current_arg_values)[index]
          else
            raise UndefinedFunctionError
          end
        end
      end
```

We alias defP to dẹf. Note that isn't a dead pixel.